requirements.txt
# Django
Django==4.2.16
djangorestframework==3.15.2
django-filter==24.3
django-cors-headers==4.4.0
django-environ==0.11.2
drf-spectacular==0.27.2

# Database
psycopg2-binary==2.9.9
pymongo==4.8.0
django-celery-beat==2.6.0
django-celery-results==2.5.1

# Celery
celery==5.4.0
redis==5.0.8

# AWS S3
boto3==1.35.0
django-storages==1.14.4

# Scraping
playwright==1.47.0
beautifulsoup4==4.12.3
lxml==5.3.0

# PDF processing
pdfplumber==0.11.4
pypdf==4.3.1

# HTTP
httpx==0.27.2
requests==2.32.3

# Notifications
python-telegram-bot==21.5

# Utils
django-extensions==3.2.3
Pillow==10.4.0
gunicorn==23.0.0
whitenoise==6.7.0
python-dateutil==2.9.0
